tests/wpt/css-tests/css-transforms-1_dev/xhtml1print/transform_translate_neg.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSS Transforms API Test: transform translate with negative translation value</title>
    <link href="http://www.intel.com" rel="author" title="Intel" />
    <link href="http://www.w3.org/TR/css-transforms-1/#transform-property" rel="help" />
    <link href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="Check if transform supports translate(-1px, -1px)" name="assert" /> 
    <script src="/resources/testharness.js" type="text/javascript"></script>
    <script src="/resources/testharnessreport.js" type="text/javascript"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
    test(function() {
        document.getElementById("test").style.transform = "translate(-1px, -1px)";
        var value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText;
        assert_equals(value, "translate(-1px, -1px)")
    }, "transform_translate_-1px_-1px");
    </script>    


</body></html>
